+++ src/main/kotlin/com/dsoftware/ghtoolbar/api/model/WorkflowRunJobModel.kt
@@ -1,3 +1,6 @@
+import com.fasterxml.jackson.annotation.JsonFormat
+import java.util.Date
+
 data class WorkflowRunJobs(
     val total_count: Int,
     val jobs: List<WorkflowRunJob>
@@ -29,46 +32,48 @@ data class WorkflowRunJobs(
 data class WorkflowRunJob(
 
     /* The id of the job. */
-    val id: kotlin.Int,
+    val id: Long,
     /* The id of the associated workflow run. */
-    val runId: kotlin.Int,
-    val runUrl: kotlin.String,
+    val runId: Long,
+    val runUrl: String,
     /* Attempt number of the associated workflow run, 1 for first attempt and higher if the workflow was re-run. */
-    val runAttempt: kotlin.Int? = null,
-    val nodeId: kotlin.String,
+    val runAttempt: Int? = null,
+    val nodeId: String,
     /* The SHA of the commit that is being run. */
-    val headSha: kotlin.String,
-    val url: kotlin.String,
-    val htmlUrl: kotlin.String,
+    val headSha: String,
+    val url: String,
+    val htmlUrl: String,
     /* The phase of the lifecycle that the job is currently in. */
     val status: WorkflowRunJob.Status,
     /* The outcome of the job. */
-    val conclusion: kotlin.String,
+    val conclusion: String,
     /* The time that the job started, in ISO 8601 format. */
-    val startedAt: java.time.LocalDateTime,
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
+    val startedAt: Date,
     /* The time that the job finished, in ISO 8601 format. */
-    val completedAt: java.time.LocalDateTime,
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'")
+    val completedAt: Date,
     /* The name of the job. */
-    val name: kotlin.String,
+    val name: String,
     /* Steps in this job. */
-    val steps: kotlin.Array<WorkflowRunJobSteps>? = null,
-    val checkRunUrl: kotlin.String,
+    val steps: Array<WorkflowRunJobSteps>? = null,
+    val checkRunUrl: String,
     /* Labels for the workflow job. Specified by the \"runs_on\" attribute in the action's workflow file. */
-    val labels: kotlin.Array<kotlin.String>,
+    val labels: Array<String>,
     /* The ID of the runner to which this job has been assigned. (If a runner hasn't yet been assigned, this will be null.) */
-    val runnerId: kotlin.Int,
+    val runnerId: Long,
     /* The name of the runner to which this job has been assigned. (If a runner hasn't yet been assigned, this will be null.) */
-    val runnerName: kotlin.String,
+    val runnerName: String?,
     /* The ID of the runner group to which this job has been assigned. (If a runner hasn't yet been assigned, this will be null.) */
-    val runnerGroupId: kotlin.Int,
+    val runnerGroupId: Long?,
     /* The name of the runner group to which this job has been assigned. (If a runner hasn't yet been assigned, this will be null.) */
-    val runnerGroupName: kotlin.String
+    val runnerGroupName: String?
 ) {
     /**
      * The phase of the lifecycle that the job is currently in.
      * Values: QUEUED,INPROGRESS,COMPLETED
      */
-    enum class Status(val value: kotlin.String) {
+    enum class Status(val value: String) {
         QUEUED("queued"),
         INPROGRESS("in_progress"),
         COMPLETED("completed");
@@ -89,20 +94,22 @@ data class WorkflowRunJobSteps(
     /* The phase of the lifecycle that the job is currently in. */
     val status: WorkflowRunJobSteps.Status,
     /* The outcome of the job. */
-    val conclusion: kotlin.String,
+    val conclusion: String,
     /* The name of the job. */
-    val name: kotlin.String,
-    val number: kotlin.Int,
+    val name: String,
+    val number: Int,
     /* The time that the step started, in ISO 8601 format. */
-    val startedAt: java.time.LocalDateTime? = null,
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSX")
+    val startedAt: Date? = null,
     /* The time that the job finished, in ISO 8601 format. */
-    val completedAt: java.time.LocalDateTime? = null
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSX")
+    val completedAt: Date? = null
 ) {
     /**
      * The phase of the lifecycle that the job is currently in.
      * Values: QUEUED,INPROGRESS,COMPLETED
      */
-    enum class Status(val value: kotlin.String) {
+    enum class Status(val value: String) {
         QUEUED("queued"),
         INPROGRESS("in_progress"),
         COMPLETED("completed");
