+++ src/main/kotlin/com/dsoftware/ghtoolbar/ui/WorkflowRunListLoaderPanel.kt
@@ -2,36 +2,210 @@ package com.dsoftware.ghtoolbar.ui.wfpanel
 
 
 import com.dsoftware.ghtoolbar.api.model.GitHubWorkflowRun
-import com.dsoftware.ghtoolbar.workflow.LoadingErrorHandler
+import com.dsoftware.ghtoolbar.ui.LoadingErrorHandler
 import com.dsoftware.ghtoolbar.workflow.WorkflowRunListSelectionHolder
-import com.dsoftware.ghtoolbar.workflow.data.WorkflowRunDataContext
-import com.dsoftware.ghtoolbar.workflow.data.WorkflowRunListLoader
+import com.dsoftware.ghtoolbar.data.WorkflowRunListLoader
+import com.dsoftware.ghtoolbar.ui.Icons
+import com.dsoftware.ghtoolbar.workflow.WorkflowRunDataContext
+import com.dsoftware.ghtoolbar.actions.ActionKeys
+import com.intellij.icons.AllIcons
+import com.intellij.ide.CopyProvider
 import com.intellij.ide.actions.RefreshAction
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.*
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.logger
+import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.progress.util.ProgressWindow
 import com.intellij.openapi.util.Disposer
 import com.intellij.ui.*
-import com.intellij.util.ui.ComponentWithEmptyText
-import com.intellij.util.ui.JBUI
-import com.intellij.util.ui.StatusText
+import com.intellij.ui.components.JBLabel
+import com.intellij.ui.components.JBList
+import com.intellij.util.text.DateFormatUtil
+import com.intellij.util.ui.*
 import com.intellij.util.ui.components.BorderLayoutPanel
 import com.intellij.vcs.log.ui.frame.ProgressStripe
+import net.miginfocom.layout.CC
+import net.miginfocom.layout.LC
+import net.miginfocom.swing.MigLayout
 import org.jetbrains.plugins.github.exceptions.GithubStatusCodeException
 import org.jetbrains.plugins.github.ui.HtmlInfoPanel
 import java.awt.BorderLayout
+import java.awt.Component
 import java.awt.event.ActionEvent
 import java.awt.event.FocusEvent
 import java.awt.event.FocusListener
-import javax.swing.JComponent
-import javax.swing.JPanel
-import javax.swing.ScrollPaneConstants
+import java.awt.event.MouseEvent
+import java.time.LocalDateTime
+import java.time.ZoneId
+import java.time.ZoneOffset
+import java.util.Date
+import javax.swing.*
 import javax.swing.event.ListDataEvent
 import javax.swing.event.ListDataListener
 import javax.swing.event.ListSelectionEvent
 
+class WorkflowRunList(model: ListModel<GitHubWorkflowRun>) : JBList<GitHubWorkflowRun>(model), DataProvider,
+    CopyProvider {
+
+    init {
+        selectionModel.selectionMode = ListSelectionModel.SINGLE_SELECTION
+
+        val renderer = WorkflowRunsListCellRenderer()
+        cellRenderer = renderer
+        putClientProperty(UIUtil.NOT_IN_HIERARCHY_COMPONENTS, listOf(renderer))
+
+        ScrollingUtil.installActions(this)
+    }
+
+    override fun getToolTipText(event: MouseEvent): String? {
+        val childComponent = ListUtil.getDeepestRendererChildComponentAt(this, event.point)
+        if (childComponent !is JComponent) return null
+        return childComponent.toolTipText
+    }
+
+    override fun getData(dataId: String): Any? = when {
+        PlatformDataKeys.COPY_PROVIDER.`is`(dataId) -> this
+        ActionKeys.SELECTED_WORKFLOW_RUN.`is`(dataId) -> selectedValue
+        else -> null
+    }
+
+    private inner class WorkflowRunsListCellRenderer : ListCellRenderer<GitHubWorkflowRun>, JPanel() {
+
+        private val stateIcon = JLabel()
+        private val title = JLabel()
+        private val info = JLabel()
+        private val labels = JPanel().apply {
+            layout = BoxLayout(this, BoxLayout.X_AXIS)
+        }
+        private val assignees = JPanel().apply {
+            layout = BoxLayout(this, BoxLayout.X_AXIS)
+        }
+
+        init {
+            border = JBUI.Borders.empty(5, 8)
+
+            layout = MigLayout(
+                LC().gridGap("0", "0")
+                    .insets("0", "0", "0", "0")
+                    .fillX()
+            )
+
+            val gapAfter = "${JBUI.scale(5)}px"
+            add(
+                stateIcon, CC()
+                    .gapAfter(gapAfter)
+            )
+            add(
+                title, CC()
+                    .growX()
+                    .pushX()
+                    .minWidth("pref/2px")
+            )
+            add(
+                labels, CC()
+                    .minWidth("pref/2px")
+                    .alignX("right")
+                    .wrap()
+            )
+            add(
+                info, CC()
+                    .minWidth("pref/2px")
+                    .skip(1)
+                    .spanX(3)
+            )
+        }
+
+        override fun getListCellRendererComponent(
+            list: JList<out GitHubWorkflowRun>,
+            ghWorkflowRun: GitHubWorkflowRun,
+            index: Int,
+            isSelected: Boolean,
+            cellHasFocus: Boolean
+        ): Component {
+            UIUtil.setBackgroundRecursively(this, ListUiUtil.WithTallRow.background(list, isSelected, list.hasFocus()))
+            val primaryTextColor = ListUiUtil.WithTallRow.foreground(isSelected, list.hasFocus())
+            val secondaryTextColor = ListUiUtil.WithTallRow.secondaryForeground(list, isSelected)
+
+            stateIcon.icon = ghWorkflowRunIcon(ghWorkflowRun)
+            title.apply {
+                text = ghWorkflowRun.head_commit.message
+                foreground = primaryTextColor
+            }
+
+            info.apply {
+                text = ghWorkflowRunInfo(ghWorkflowRun)
+                foreground = secondaryTextColor
+            }
+            labels.apply {
+                removeAll()
+                add(JBLabel(" ${ghWorkflowRun.head_branch} ", UIUtil.ComponentStyle.SMALL).apply {
+                    foreground = JBColor(ColorUtil.softer(secondaryTextColor), ColorUtil.softer(secondaryTextColor))
+                })
+                add(Box.createRigidArea(JBDimension(4, 0)))
+            }
+            return this
+        }
+    }
+
+    companion object {
+        private val LOG = thisLogger()
+        fun ghWorkflowRunIcon(ghWorkflowRun: GitHubWorkflowRun): Icon {
+            return when (ghWorkflowRun.status) {
+                "completed" -> {
+                    when (ghWorkflowRun.conclusion) {
+                        "success" -> AllIcons.Actions.Commit
+                        "failure" -> Icons.X
+                        else -> Icons.PrimitiveDot
+                    }
+                }
+                "queued" -> Icons.PrimitiveDot
+                "in progress" -> Icons.PrimitiveDot
+                "neutral" -> Icons.PrimitiveDot
+                "success" -> AllIcons.Actions.Commit
+                "failure" -> Icons.X
+                "cancelled" -> Icons.X
+                "action required" -> Icons.Watch
+                "timed out" -> Icons.Watch
+                "skipped" -> Icons.X
+                "stale" -> Icons.Watch
+                else -> Icons.PrimitiveDot
+            }
+        }
+
+        fun ghWorkflowRunInfo(ghWorkflowRun: GitHubWorkflowRun): String {
+            val updatedAtLabel =
+                if (ghWorkflowRun.updated_at == null) "Unknown"
+                else makeTimePretty(ghWorkflowRun.updated_at)
+
+            var action = "pushed by"
+            if (ghWorkflowRun.event == "release") {
+                action = "created by"
+            }
+            return "${ghWorkflowRun.name} #${ghWorkflowRun.run_number}: " +
+                "$action ${ghWorkflowRun.head_commit.author.name} " +
+                "on $updatedAtLabel"
+        }
+
+        fun makeTimePretty(date: Date): String {
+            val localDateTime = LocalDateTime.ofInstant(date.toInstant(), ZoneId.systemDefault())
+            val zonedDateTime = localDateTime.atZone(ZoneOffset.UTC)
+            return DateFormatUtil.formatPrettyDateTime(zonedDateTime.toInstant().toEpochMilli())
+        }
+    }
+
+    override fun performCopy(dataContext: DataContext) {
+        TODO("Not yet implemented")
+    }
+
+    override fun isCopyEnabled(dataContext: DataContext): Boolean {
+        return false
+    }
+
+    override fun isCopyVisible(dataContext: DataContext): Boolean {
+        return false
+    }
+}
 internal class WorkflowRunListLoaderPanel(
     disposable: Disposable,
     private val runListLoader: WorkflowRunListLoader,
@@ -263,7 +437,7 @@ internal class WorkflowRunListLoaderPanel(
             listSelectionHolder: WorkflowRunListSelectionHolder,
             disposable: Disposable,
         ): JComponent {
-            val list = WorkflowRunList(context.listModel).apply {
+            val list = WorkflowRunList(context.runsListModel).apply {
                 emptyText.clear()
             }.also {
                 it.addFocusListener(object : FocusListener {
@@ -280,7 +454,7 @@ internal class WorkflowRunListLoaderPanel(
 
             val listReloadAction = actionManager.getAction("Github.Workflow.List.Reload") as RefreshAction
 
-            return WorkflowRunListLoaderPanel(disposable, context.listLoader, listReloadAction, list)
+            return WorkflowRunListLoaderPanel(disposable, context.runsListLoader, listReloadAction, list)
         }
 
         private fun getLoadingErrorText(error: Throwable, newLineSeparator: String = "\n"): String {
